+++ planet-mitchell/src/Optic/Prism.hs
@@ -8,6 +8,10 @@ module Optic.Prism
   , prism'
   , is
   , only
+    -- * AsAny
+  , AsAny(..)
+    -- * AsSubtype
+  , AsSubtype(..)
     -- * Cons
   , Cons(..)
   , (<|)
@@ -29,9 +33,11 @@ module Optic.Prism
   , pattern Empty
   ) where
 
-import Control.Lens.Cons   (pattern (:<), pattern (:>), Cons(_Cons),
-                            Snoc(_Snoc), cons, snoc, uncons, unsnoc, (<|),
-                            _head, _init, _last, _tail, (|>))
-import Control.Lens.Empty  (AsEmpty(_Empty), pattern Empty)
-import Control.Lens.Extras (is)
-import Control.Lens.Prism  (Prism, Prism', only, prism, prism')
+import Control.Lens.Cons         (pattern (:<), pattern (:>), Cons(_Cons),
+                                  Snoc(_Snoc), cons, snoc, uncons, unsnoc, (<|),
+                                  _head, _init, _last, _tail, (|>))
+import Control.Lens.Empty        (AsEmpty(_Empty), pattern Empty)
+import Control.Lens.Extras       (is)
+import Control.Lens.Prism        (Prism, Prism', only, prism, prism')
+import Data.Generics.Sum.Any     (AsAny(_As))
+import Data.Generics.Sum.Subtype (AsSubtype(_Sub))